requirements.txt
colorama==0.3.9
jsonschema==2.6.0
prettytable==0.7.2
python-binance==0.7.0
python-dateutil==2.7.3
SQLAlchemy==1.2.11
